Added Firebase functionality to EditItemFragment.java

+++ app/src/main/java/com/b07project/application/AddItemFragment.java
@@ -88,7 +88,7 @@ public class AddItemFragment extends Fragment {
                 String name = itemName.getText().toString().trim();
                 String description = itemDescription.getText().toString().trim();
                 String specification = itemSpecifications.getText().toString().trim();
-                Double price = Double.valueOf(itemPrice.getText().toString().trim());
+                Float price = Float.valueOf(itemPrice.getText().toString().trim());
 
                 Item item = new Item(name, description, price, mParam1, specification);
                 item.saveItem();
